♻️Refactor : 채팅 메시지 시간 포맷팅
