--- src/main/java/roomit/main/domain/chat/chatmessage/dto/ChatMessageRequest.java
@@ -1,14 +1,0 @@
-package roomit.main.domain.chat.chatmessage.dto;
-
-import lombok.Builder;
-
-import java.time.LocalDateTime;
-
-public record ChatMessageRequest(
-        Long roomId,
-        String sender,
-        String content,
-        LocalDateTime timestamp,
-        String senderType // business or member
-        ) {
-}
