+++ src/main/java/roomit/main/domain/chat/chatroom/dto/response/ChatRoomBusinessResponse.java
@@ -1,5 +1,6 @@
 package roomit.main.domain.chat.chatroom.dto.response;
 
+import com.fasterxml.jackson.annotation.JsonFormat;
 import roomit.main.domain.chat.chatmessage.entity.ChatMessage;
 import roomit.main.domain.chat.chatroom.entity.ChatRoom;
 
@@ -11,7 +12,7 @@ public record ChatRoomBusinessResponse(
         String memberNickname,
         String messageContent,
         Boolean isRead,
-        LocalDateTime updatedAt
+        @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd'T'HH:mm:ss") LocalDateTime updatedAt
 ) implements ChatRoomResponse {
 
     public ChatRoomBusinessResponse(ChatRoom chatRoom, ChatMessage chatMessage) {
