+++ src/main/java/roomit/main/domain/chat/chatroom/dto/response/ChatRoomMemberResponse.java
@@ -1,5 +1,6 @@
 package roomit.main.domain.chat.chatroom.dto.response;
 
+import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.Builder;
 import roomit.main.domain.chat.chatmessage.entity.ChatMessage;
 import roomit.main.domain.chat.chatroom.entity.ChatRoom;
@@ -12,7 +13,7 @@ public record ChatRoomMemberResponse(
         String workplaceName,
         String messageContent,
         Boolean isRead,
-        LocalDateTime updatedAt
+        @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd'T'HH:mm:ss") LocalDateTime updatedAt
 ) implements ChatRoomResponse {
 
     public ChatRoomMemberResponse(ChatRoom chatRoom, ChatMessage chatMessage, String workplaceName) {
