+++ src/main/java/roomit/main/domain/chat/chatroom/dto/response/ChatRoomResponse.java
@@ -1,5 +1,7 @@
 package roomit.main.domain.chat.chatroom.dto.response;
 
+import com.fasterxml.jackson.annotation.JsonFormat;
+
 import java.time.LocalDateTime;
 
 public interface ChatRoomResponse {
@@ -8,5 +10,5 @@ public interface ChatRoomResponse {
     String name();
     String messageContent();
     Boolean isRead();
-    LocalDateTime updatedAt();
+    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd'T'HH:mm:ss") LocalDateTime updatedAt();
 }
